pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cat-watcher"
version = "0.1.0"
description = "ML-powered cat behavior detection and identification system"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "jeefy" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "aiomqtt>=2.0.0",
    "Pillow>=10.0.0",
    "numpy>=1.26.0",
    "pyyaml>=6.0.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
inference = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "ultralytics>=8.1.0",
    "onnxruntime-gpu>=1.16.0",
    "timm>=0.9.0",
    "opencv-python>=4.8.0",
    "scipy>=1.11.0",
]
training = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "ultralytics>=8.1.0",
    "timm>=0.9.0",
    "albumentations>=1.3.0",
    "opencv-python>=4.8.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "scipy>=1.11.0",
]
labeling = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "jinja2>=3.1.0",
    "aiosqlite>=0.19.0",
    "python-multipart>=0.0.6",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "respx>=0.20.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]
all = [
    "cat-watcher[inference,training,labeling,dev]",
]

[project.scripts]
cat-watcher = "cat_watcher.cli:main"
cat-watcher-collect = "cat_watcher.collection.collector:main"
cat-watcher-label = "cat_watcher.labeling.app:main"
cat-watcher-inference = "cat_watcher.inference.service:main"
cat-watcher-inference-api = "cat_watcher.inference.app:main"

[project.urls]
Homepage = "https://github.com/jeefy/cat-watcher"
Repository = "https://github.com/jeefy/cat-watcher"

[tool.hatch.build.targets.wheel]
packages = ["src/cat_watcher"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["cat_watcher"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]
mypy_path = "src"
explicit_package_bases = true

[[tool.mypy.overrides]]
module = [
    "ultralytics.*",
    "cv2.*",
    "albumentations.*",
    "timm.*",
    "aiomqtt.*",
]
ignore_missing_imports = true

[tool.coverage.run]
source = ["src/cat_watcher"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
